+++ string_calculator/webapp/src/main/java/pl/tzch96/string_calculator/webapp/CalculatorController.java
@@ -1,0 +1,16 @@
+package pl.tzch96.string_calculator.webapp;
+
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.ui.Model;
+
+@RestController
+public class CalculatorController {
+    String appName = "string_calculator";
+
+    @GetMapping("/")
+    public String indexPage(Model model) {
+        model.addAttribute("appName", appName);
+        return "index";
+    }
+}