+++ string_calculator/webapp/src/main/java/pl/tzch96/string_calculator/webapp/CalculatorWebapp.java
@@ -1,0 +1,13 @@
+package pl.tzch96.string_calculator.webapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CalculatorWebapp {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CalculatorWebapp.class, args);
+    }
+
+}