+++ string_calculator/webapp/src/test/java/pl/tzch96/string_calculator/webapp/WebappApplicationTests.java
@@ -1,0 +1,13 @@
+package pl.tzch96.string_calculator.webapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WebappApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}